pyproject.toml
[project]
name = "field"
version = "0.1.0"
description = "reddit field stuff"
requires-python = ">=3.10"
dependencies = [
    "websockets",
    "gql[all]<4",
    "redis",
    "hiredis",
    "requests",
    "httpx[http2]",
    "lru-dict",
    "aiofiles",
    "orjson",
    "sqlalchemy[asyncio]<3"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit"
]
